% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\docType{data}
\name{blocks}
\alias{blocks}
\alias{clinic}
\title{Blocks}
\format{
An object of class \code{list} of length 1.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 213 rows and 118 columns.
}
\usage{
data(blocks)

data(clinic)
}
\description{
Tables of omics and clinic from the ImmunAID project (H2020) for
Auto-Inflammatory Diseases (AID).
}
\keyword{datasets}
