% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAnalysis-package.R
\docType{package}
\name{multiAnalysis-package}
\alias{multiAnalysis}
\alias{multiAnalysis-package}
\title{multiAnalysis: Multi-omics data analysis for Auto-Inflammatory Diseases}
\description{
Multi-omics data analysis for the ImmunAID project (H2020):
    identify potential biomarkers for Auto-Inflammatory Diseases (AID).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecamenen/multiAnalysis}
  \item Report bugs at \url{https://github.com/ecamenen/multiAnalysis/issues}
}

}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
